fix: move content order to flight, wrong place
